backend/requirements.txt
Flask==2.1.2
Werkzeug==2.0.3
mysql-connector-python==8.0.33
Flask-Cors==3.0.10
